Add v1 integration; refactor v2 common to use common.go configuration
