+++ pkg/integrations/ebpf_exporter/ebpf_test.go
@@ -1,0 +1,7 @@
+package ebpf
+
+import "testing"
+
+func TestHelloWorld(t *testing.T) {
+	// t.Fatal("not implemented")
+}
